% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratVisProExample.R
\name{SeuratVisProExample}
\alias{SeuratVisProExample}
\title{Generate \code{Seurat} object for demonstrations}
\usage{
SeuratVisProExample(
  n_cells = 300,
  n_genes = 1000,
  n_clusters = 10,
  seed = 123,
  genes_mt = "^MT-",
  neighbor_dims = 10,
  cluster_res = 0.5,
  umap_dims = 10,
  spatial = FALSE
)
}
\arguments{
\item{n_cells}{Number of cells. Default: \code{300}.}

\item{n_genes}{Number of genes. Default: \code{1000}.}

\item{n_clusters}{Number of clusters. Default: \code{10}.}

\item{seed}{Random seed for reproducibility. Default: \code{123}.}

\item{genes_mt}{Regex for mitochondrial genes. Default: \code{"^MT-"}.}

\item{neighbor_dims}{Neighbor graph dimensions used in \code{FindNeighbors}. Default: \code{10}.}

\item{cluster_res}{Cluster resolution used in \code{FindClusters}. Default: \code{0.5}.}

\item{umap_dims}{UMAP dimensions used in \code{RunUMAP}. Default: \code{10}.}

\item{spatial}{Whether to add synthetic spatial coordinates \code{x},\code{y} to \code{meta.data}. Default: \code{FALSE}.}
}
\value{
A \code{Seurat} object with basic preprocessing steps applied.
}
\description{
Generate a \code{Seurat} object with basic preprocessing (DefaultAssay, PercentageFeatureSet, NormalizeData, FindVariableFeatures, ScaleData, RunPCA, FindNeighbors, FindClusters, RunUMAP).
}
\examples{
obj <- SeuratVisProExample(
    n_cells = 300,
    n_genes = 1000,
    n_clusters = 10,
    seed = 123,
    genes_mt = "^MT-",
    neighbor_dims = 10,
    cluster_res = 0.5,
    umap_dims = 10,
    spatial = FALSE)

Seurat::DimPlot(obj, group.by = "cluster")

}
\author{
benben-miao
}
