% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisBatchAlign.R
\name{VisBatchAlign}
\alias{VisBatchAlign}
\title{Batch mixing diagnostics}
\usage{
VisBatchAlign(
  object,
  batch,
  reduction = "pca",
  dims = 1:10,
  k = 20,
  palette = "C",
  violin_width = 0.8,
  violin_alpha = 0.3,
  box_width = 0.3,
  box_alpha = 0.5
)
}
\arguments{
\item{object}{A \code{Seurat} object.}

\item{batch}{Column name in \code{meta.data} indicating batch.}

\item{reduction}{Dimensional reduction to use for kNN (e.g., 'pca' or 'umap').}

\item{dims}{Dimensions to use.}

\item{k}{Number of neighbors.}

\item{palette}{Viridis palette option.}

\item{violin_width}{Violin width.}

\item{violin_alpha}{Violin alpha.}

\item{box_width}{Box width.}

\item{box_alpha}{Box alpha.}
}
\value{
A list with summary data and plot comparing neighbor batch composition.
}
\description{
Quantify and visualize batch mixing in a KNN graph before/after integration.
}
\examples{
obj <- SeuratVisProExample(
    n_cells = 300,
    n_genes = 1000,
    n_clusters = 10,
    seed = 123,
    genes_mt = "^MT-",
    neighbor_dims = 10,
    cluster_res = 0.5,
    umap_dims = 10,
    spatial = FALSE)

obj$batch <- sample(
  c('A','B'),
  ncol(obj),
  replace = TRUE)

res <- VisBatchAlign(
  obj,
  batch = 'batch',
  reduction = 'pca',
  dims = 1:10,
  k = 20,
  palette = "C",
  violin_width = 0.8,
  violin_alpha = 0.3,
  box_width = 0.3,
  box_alpha = 0.5)

res$plot
head(res$summary)

}
\author{
benben-miao
}
