% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisClusterTree.R
\name{VisClusterTree}
\alias{VisClusterTree}
\title{Cluster dendrogram visualization}
\usage{
VisClusterTree(
  object,
  assay = NULL,
  group.by = "seurat_clusters",
  dist.metric = "euclidean",
  linkage = "complete",
  show_heatmap = TRUE,
  palette = "C",
  tile_alpha = 0.8
)
}
\arguments{
\item{object}{A \code{Seurat} object; required.}

\item{assay}{Assay name. Default: \code{Seurat::DefaultAssay(object)}.}

\item{group.by}{Metadata column for cluster identity. Default: \code{"seurat_clusters"}.}

\item{dist.metric}{Distance metric: \code{'euclidean'} or \code{'correlation'}. Default: \code{'euclidean'}.}

\item{linkage}{Linkage method (e.g., \code{'complete'}, \code{'average'}, \code{'ward.D2'}). Default: \code{'complete'}.}

\item{show_heatmap}{Append pairwise similarity heatmap under the tree. Default: \code{TRUE}.}

\item{palette}{Viridis palette option for color/fill. Default: \code{"C"}.}

\item{tile_alpha}{Tile alpha. Default: \code{0.8}.}
}
\value{
A ggplot dendrogram (or patchwork with heatmap).
}
\description{
Build a hierarchical tree of clusters using average expression profiles.
}
\examples{
obj <- SeuratVisProExample(
    n_cells = 300,
    n_genes = 1000,
    n_clusters = 10,
    seed = 123,
    genes_mt = "^MT-",
    neighbor_dims = 10,
    cluster_res = 0.5,
    umap_dims = 10,
    spatial = FALSE)

p <- VisClusterTree(
  obj,
  assay = NULL,
  group.by = "seurat_clusters",
  dist.metric = "euclidean",
  linkage = "complete",
  show_heatmap = TRUE,
  palette = "C",
  tile_alpha = 0.8)
p

}
\author{
benben-miao
}
