% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisGeneTrend.R
\name{VisGeneTrend}
\alias{VisGeneTrend}
\title{Gene expression trends along pseudotime or groups}
\usage{
VisGeneTrend(
  object,
  features,
  by = "pseudotime",
  reduction = "umap",
  dims = 1:2,
  smooth.method = "loess",
  palette = "C"
)
}
\arguments{
\item{object}{A \code{Seurat} object with UMAP or PCA.}

\item{features}{Character vector of gene names.}

\item{by}{Either 'pseudotime' or a metadata column to group by.}

\item{reduction}{Reduction to use for pseudotime ('umap' or 'pca').}

\item{dims}{Dimensions used for pseudotime ranking.}

\item{smooth.method}{Smoothing method ('loess' or 'gam').}

\item{palette}{Viridis palette option.}
}
\value{
A ggplot with smoothed curves.
}
\description{
Plot smoothed trends for selected genes across pseudotime approximated from embeddings or grouped by clusters.
}
\examples{
obj <- SeuratVisProExample()

p <- VisGeneTrend(
  obj,
  features = c("G10", "G20", "G30"),
  by = "pseudotime",
  reduction = "umap",
  dims = 1:2,
  smooth.method = "loess",
  palette = "C")
p

}
\author{
benben-miao
}
