% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisLocalMoran.R
\name{VisLocalMoran}
\alias{VisLocalMoran}
\title{Local Moran's I hotspot for gene expression}
\usage{
VisLocalMoran(
  object,
  gene,
  reduction = "umap",
  k = 15,
  palette = "C",
  point_size = 2,
  point_alpha = 0.8
)
}
\arguments{
\item{object}{A \code{Seurat} object; UMAP/PCA is computed when absent.}

\item{gene}{Gene name; must exist in assay data.}

\item{reduction}{Reduction name (\code{'umap'} or \code{'pca'}). Default: \code{'umap'}.}

\item{k}{Number of nearest neighbors. Default: \code{15}.}

\item{palette}{Viridis palette option for color/fill. Default: \code{"C"}.}

\item{point_size}{Point size. Default: \code{2}.}

\item{point_alpha}{Point alpha. Default: \code{0.8}.}
}
\value{
A ggplot scatter of local Moran's I values.
}
\description{
Compute local Moran's I on UMAP/PCA neighborhoods for a given gene to detect spatial autocorrelation hotspots.
}
\examples{
obj <- SeuratVisProExample(
    n_cells = 300,
    n_genes = 1000,
    n_clusters = 10,
    seed = 123,
    genes_mt = "^MT-",
    neighbor_dims = 10,
    cluster_res = 0.5,
    umap_dims = 10,
    spatial = FALSE)

p <- VisLocalMoran(
   obj,
   gene = 'G10',
   reduction = "umap",
   k = 15,
   palette = "C",
   point_size = 2,
   point_alpha = 0.8)
p

}
\author{
benben-miao
}
