% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratVisProExample.R
\name{SeuratVisProExample}
\alias{SeuratVisProExample}
\title{Generate \code{Seurat} object for demonstrations}
\usage{
SeuratVisProExample(
  n_cells = 300,
  n_genes = 1000,
  n_clusters = 10,
  seed = 123,
  genes_mt = "^MT-",
  neighbor_dims = 10,
  cluster_res = 0.5,
  umap_dims = 10,
  spatial = FALSE
)
}
\arguments{
\item{n_cells}{Number of cells.}

\item{n_genes}{Number of genes.}

\item{n_clusters}{Number of simulated clusters.}

\item{seed}{Random seed for reproducibility.}

\item{spatial}{If TRUE, adds synthetic spatial coordinates 'x','y' to meta.data.}
}
\value{
A \code{Seurat} object with basic preprocessing (DefaultAssay, PercentageFeatureSet, NormalizeData, FindVariableFeatures, ScaleData, RunPCA, FindNeighbors, FindClusters, RunUMAP).
}
\description{
Generate a \code{Seurat} object with basic preprocessing (DefaultAssay, PercentageFeatureSet, NormalizeData, FindVariableFeatures, ScaleData, RunPCA, FindNeighbors, FindClusters, RunUMAP).
}
\examples{
obj <- SeuratVisProExample(
    n_cells = 300,
    n_genes = 1000,
    n_clusters = 10,
    seed = 123,
    spatial = FALSE)

Seurat::DimPlot(obj, group.by = "cluster")

}
\author{
benben-miao
}
