% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisBatchAlign.R
\name{VisBatchAlign}
\alias{VisBatchAlign}
\title{Batch mixing diagnostics}
\usage{
VisBatchAlign(
  object,
  batch,
  reduction = "pca",
  dims = 1:10,
  k = 20,
  palette = "C"
)
}
\arguments{
\item{object}{A \code{Seurat} object.}

\item{batch}{Column name in \code{meta.data} indicating batch.}

\item{reduction}{Dimensional reduction to use for kNN (e.g., 'pca' or 'umap').}

\item{dims}{Dimensions to use.}

\item{k}{Number of neighbors.}

\item{palette}{Viridis palette option.}
}
\value{
A list with summary data and plot comparing neighbor batch composition.
}
\description{
Quantify and visualize batch mixing in a KNN graph before/after integration.
}
\examples{
obj <- SeuratVisProExample()

obj$batch <- sample(c('A','B'), ncol(obj), replace = TRUE)

res <- VisBatchAlign(obj, batch = 'batch', reduction = 'pca', dims = 1:10, k = 20, palette = "C")
res$plot
head(res$summary)

}
\author{
benben-miao
}
