% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisBatchAlign.R
\name{VisBatchAlign}
\alias{VisBatchAlign}
\title{Batch mixing diagnostics}
\usage{
VisBatchAlign(
  object,
  batch,
  reduction = "pca",
  dims = 1:10,
  k = 20,
  palette = "C",
  violin_width = 0.8,
  violin_alpha = 0.3,
  box_width = 0.3,
  box_alpha = 0.5
)
}
\arguments{
\item{object}{A \code{Seurat} object; required.}

\item{batch}{Column in \code{object@meta.data} indicating batch; required.}

\item{reduction}{Reduction for kNN computation, \code{'pca'} or \code{'umap'}. Default: \code{'pca'}.}

\item{dims}{Integer vector of dimensions used. Default: \code{1:10}.}

\item{k}{Number of neighbors. Default: \code{20}.}

\item{palette}{Viridis palette option for color/fill. Default: \code{"C"}.}

\item{violin_width}{Violin width. Default: \code{0.8}.}

\item{violin_alpha}{Violin alpha. Default: \code{0.3}.}

\item{box_width}{Box width. Default: \code{0.3}.}

\item{box_alpha}{Box alpha. Default: \code{0.5}.}
}
\value{
A list with \code{summary} (per-cell mixing proportion) and \code{plot} (violin+box).
}
\description{
Quantify and visualize batch mixing in a KNN graph before/after integration.
}
\examples{
obj <- SeuratVisProExample(
    n_cells = 300,
    n_genes = 1000,
    n_clusters = 10,
    seed = 123,
    genes_mt = "^MT-",
    neighbor_dims = 10,
    cluster_res = 0.5,
    umap_dims = 10,
    spatial = FALSE)

obj$batch <- sample(
  c('A','B'),
  ncol(obj),
  replace = TRUE)

res <- VisBatchAlign(
  obj,
  batch = 'batch',
  reduction = 'pca',
  dims = 1:10,
  k = 20,
  palette = "C",
  violin_width = 0.8,
  violin_alpha = 0.3,
  box_width = 0.3,
  box_alpha = 0.5)

res$plot
head(res$summary)

}
\author{
benben-miao
}
