% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisCellCycle.R
\name{VisCellCycle}
\alias{VisCellCycle}
\title{Cell cycle visualization}
\usage{
VisCellCycle(object, s.genes, g2m.genes, reduction = "umap", dims = 1:10)
}
\arguments{
\item{object}{A \code{Seurat} object.}

\item{s.genes}{Character vector for S phase genes.}

\item{g2m.genes}{Character vector for G2/M phase genes.}

\item{reduction}{Reduction name for overlay.}

\item{dims}{Dimensions for neighbor graph.}
}
\value{
A list with updated object and patchwork plot.
}
\description{
Score cell cycle and visualize on embedding and distribution.
}
\examples{
obj <- SeuratVisProExample()
s.genes <- paste0('G', 1:10); g2m.genes <- paste0('G', 11:20)
res <- VisCellCycle(obj, s.genes, g2m.genes)
res$plot
}
