% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisClusterStability.R
\name{VisClusterStability}
\alias{VisClusterStability}
\title{Cluster stability across resolutions}
\usage{
VisClusterStability(
  object,
  resolution_range = seq(0.2, 1.2, by = 0.2),
  dims = 1:10,
  reps = 5,
  prop = 0.8,
  palette = "C"
)
}
\arguments{
\item{object}{A \code{Seurat} object with PCA computed.}

\item{resolution_range}{Numeric vector of resolutions, e.g., \code{seq(0.2, 1.2, by=0.2)}.}

\item{dims}{PCA dimensions to use.}

\item{reps}{Number of resampling repetitions.}

\item{prop}{Proportion of cells to sample per repetition.}

\item{palette}{Viridis palette option.}
}
\value{
A list with \code{summary} data.frame and a ggplot summarizing stability.
}
\description{
Assess clustering stability by resampling cells and varying resolution.
}
\examples{
obj <- SeuratVisProExample(
    n_cells = 300,
    n_genes = 1000,
    n_clusters = 10,
    seed = 123,
    genes_mt = "^MT-",
    neighbor_dims = 10,
    cluster_res = 0.5,
    umap_dims = 10,
    spatial = FALSE)

res <- VisClusterStability(
  obj,
  resolution_range = seq(0.2, 1.2, by = 0.2),
  dims = 1:10,
  reps = 5,
  prop = 0.8,
  palette = "C")

res$plot
head(res$summary)

}
\author{
benben-miao
}
