% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisEmbeddingContour.R
\name{VisEmbeddingContour}
\alias{VisEmbeddingContour}
\title{Embedding density contours per group}
\usage{
VisEmbeddingContour(
  object,
  group.by = "seurat_clusters",
  reduction = "umap",
  levels = 5,
  palette = "C",
  point_size = 1,
  point_alpha = 0.5,
  contour_alpha = 0.1
)
}
\arguments{
\item{object}{A \code{Seurat} object with UMAP or PCA.}

\item{group.by}{Metadata column for grouping.}

\item{reduction}{Reduction name, default 'umap'.}

\item{levels}{Number of contour levels.}

\item{palette}{Optional manual palette for groups.}

\item{point_size}{Point size.}

\item{point_alpha}{Point alpha.}

\item{contour_alpha}{Contour alpha.}
}
\value{
A ggplot with contours.
}
\description{
Overlay 2D density contours for groups on UMAP/PCA to highlight boundaries.
}
\examples{
obj <- SeuratVisProExample(
    n_cells = 300,
    n_genes = 1000,
    n_clusters = 10,
    seed = 123,
    genes_mt = "^MT-",
    neighbor_dims = 10,
    cluster_res = 0.5,
    umap_dims = 10,
    spatial = FALSE)

p <- VisEmbeddingContour(
  obj,
  group.by = "seurat_clusters",
  reduction = "umap",
  levels = 5,
  palette = "C",
  point_size = 1,
  point_alpha = 0.5,
  contour_alpha = 0.1)
p

}
\author{
benben-miao
}
