% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisGeneCoexpHive.R
\name{VisGeneCoexpHive}
\alias{VisGeneCoexpHive}
\title{Radial hive plot of gene co-expression}
\usage{
VisGeneCoexpHive(
  object,
  genes,
  reduction = "pca",
  threshold = 0.2,
  palette = "C",
  point_size = 3,
  point_alpha = 0.8,
  label_size = 3,
  curve_alpha = 0.5
)
}
\arguments{
\item{object}{A \code{Seurat} object; required.}

\item{genes}{Character vector of genes; must contain ≥3 present in the assay.}

\item{reduction}{Reduction used to position genes radially by loading (\code{'pca'}). Default: \code{'pca'}.}

\item{threshold}{Minimum absolute correlation to draw edges. Default: \code{0.2}.}

\item{palette}{Viridis palette option for color/fill. Default: \code{"C"}.}

\item{point_size}{Node point size. Default: \code{3}.}

\item{point_alpha}{Node point alpha. Default: \code{0.8}.}

\item{label_size}{Label size. Default: \code{3}.}

\item{curve_alpha}{Edge curve alpha. Default: \code{0.5}.}
}
\value{
A ggplot radial hive-like network.
}
\description{
Visualize a radial network of selected genes; edge weight reflects co-expression across cells.
}
\examples{
obj <- SeuratVisProExample(
    n_cells = 300,
    n_genes = 1000,
    n_clusters = 10,
    seed = 123,
    genes_mt = "^MT-",
    neighbor_dims = 10,
    cluster_res = 0.5,
    umap_dims = 10,
    spatial = FALSE)

p <- VisGeneCoexpHive(
  obj,
  genes = paste0("G", 1:100),
  reduction = "pca",
  threshold = 0.2,
  palette = "C",
  point_size = 3,
  point_alpha = 0.8,
  label_size = 3,
  curve_alpha = 0.5)
p

}
\author{
benben-miao
}
