% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisHexEntropy.R
\name{VisHexEntropy}
\alias{VisHexEntropy}
\title{Hex-binned entropy map on embeddings}
\usage{
VisHexEntropy(
  object,
  group.by = "seurat_clusters",
  reduction = "umap",
  bins = 30,
  palette = "C"
)
}
\arguments{
\item{object}{A \code{Seurat} object.}

\item{group.by}{Metadata column for groups.}

\item{reduction}{Reduction name, default 'umap'.}

\item{bins}{Number of bins per axis.}

\item{palette}{Palette.}
}
\value{
A ggplot heatmap of entropy.
}
\description{
Compute per-bin Shannon entropy of group composition on UMAP/PCA and visualize as a tile heatmap.
}
\examples{
obj <- SeuratVisProExample(
    n_cells = 300,
    n_genes = 1000,
    n_clusters = 10,
    seed = 123,
    genes_mt = "^MT-",
    neighbor_dims = 10,
    cluster_res = 0.5,
    umap_dims = 10,
    spatial = FALSE)

p <- VisHexEntropy(
  obj,
  group.by = "seurat_clusters",
  reduction = "umap",
  bins = 30,
  palette = "C")
p

}
\author{
benben-miao
}
