% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisLocalMoran.R
\name{VisLocalMoran}
\alias{VisLocalMoran}
\title{Local Moran's I hotspot for gene expression}
\usage{
VisLocalMoran(object, gene, reduction = "umap", k = 15)
}
\arguments{
\item{object}{A \code{Seurat} object.}

\item{gene}{Gene name.}

\item{reduction}{Reduction name, default 'umap'.}

\item{k}{Number of nearest neighbors.}
}
\value{
A ggplot scatter of local Moran's I values.
}
\description{
Compute local Moran's I on UMAP/PCA neighborhoods for a given gene to detect spatial autocorrelation hotspots.
}
\examples{
obj <- SeuratVisProExample(n_cells = 300, n_genes = 800, n_clusters = 3)
p <- VisLocalMoran(obj, gene = 'G10', k = 15)
p
}
