% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisLocalMoran.R
\name{VisLocalMoran}
\alias{VisLocalMoran}
\title{Local Moran's I hotspot for gene expression}
\usage{
VisLocalMoran(
  object,
  gene,
  reduction = "umap",
  k = 15,
  palette = "C",
  point_size = 2,
  point_alpha = 0.8
)
}
\arguments{
\item{object}{A \code{Seurat} object.}

\item{gene}{Gene name.}

\item{reduction}{Reduction name, default 'umap'.}

\item{k}{Number of nearest neighbors.}

\item{palette}{Palette.}

\item{point_size}{Point size.}

\item{point_alpha}{Point alpha.}
}
\value{
A ggplot scatter of local Moran's I values.
}
\description{
Compute local Moran's I on UMAP/PCA neighborhoods for a given gene to detect spatial autocorrelation hotspots.
}
\examples{
obj <- SeuratVisProExample(
    n_cells = 300,
    n_genes = 1000,
    n_clusters = 10,
    seed = 123,
    genes_mt = "^MT-",
    neighbor_dims = 10,
    cluster_res = 0.5,
    umap_dims = 10,
    spatial = FALSE)

p <- VisLocalMoran(
   obj,
   gene = 'G10',
   reduction = "umap",
   k = 15,
   palette = "C",
   point_size = 2,
   point_alpha = 0.8)
p

}
\author{
benben-miao
}
