% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisMarkerAtlas.R
\name{VisMarkerAtlas}
\alias{VisMarkerAtlas}
\title{Marker atlas visualization}
\usage{
VisMarkerAtlas(
  object,
  markers_top = 5,
  logfc_threshold = 0.25,
  min_percent = 0.1,
  test_method = "wilcox",
  palette = "C"
)
}
\arguments{
\item{object}{A \code{Seurat} object; required.}

\item{markers_top}{Number of top markers per cluster. Default: \code{5}.}

\item{logfc_threshold}{Log fold-change threshold for marker finding. Default: \code{0.25}.}

\item{min_percent}{Minimum percent expressed. Default: \code{0.1}.}

\item{test_method}{Differential test method, e.g., \code{'wilcox'}, \code{'t'}, \code{'LR'}. Default: \code{'wilcox'}.}

\item{palette}{Viridis palette option for color/fill. Default: \code{"C"}.}
}
\value{
A list with \code{markers} (top markers table) and \code{plot} (heatmap ggplot).
}
\description{
Find top markers per cluster and visualize as a heatmap.
}
\examples{
obj <- SeuratVisProExample(
    n_cells = 300,
    n_genes = 1000,
    n_clusters = 10,
    seed = 123,
    genes_mt = "^MT-",
    neighbor_dims = 10,
    cluster_res = 0.5,
    umap_dims = 10,
    spatial = FALSE)

res <- VisMarkerAtlas(
  obj,
  markers_top = 5,
  logfc_threshold = 0.25,
  min_percent = 0.1,
  test_method = "wilcox",
  palette = "C")

res$plot
head(res$markers)

}
\author{
benben-miao
}
