% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisMetaFeature.R
\name{VisMetaFeature}
\alias{VisMetaFeature}
\title{Module score visualization for gene sets}
\usage{
VisMetaFeature(
  object,
  feature_sets,
  group.by = "seurat_clusters",
  nbin = 24,
  min.size = 3,
  palette = "C",
  violin_width = 0.8,
  violin_alpha = 0.3,
  box_width = 0.3,
  box_alpha = 0.5
)
}
\arguments{
\item{object}{A \code{Seurat} object; required.}

\item{feature_sets}{A named list of character vectors of genes; required.}

\item{group.by}{Metadata column for grouping. Default: \code{"seurat_clusters"}.}

\item{nbin}{Number of bins for \code{Seurat::AddModuleScore}. Default: \code{24}.}

\item{min.size}{Minimum gene set size to keep. Default: \code{3}.}

\item{palette}{Viridis palette option for color/fill. Default: \code{"C"}.}

\item{violin_width}{Violin width. Default: \code{0.8}.}

\item{violin_alpha}{Violin alpha. Default: \code{0.3}.}

\item{box_width}{Box width. Default: \code{0.3}.}

\item{box_alpha}{Box alpha. Default: \code{0.5}.}
}
\value{
A list with \code{object} and \code{plot} (violin+box per gene set).
}
\description{
Compute module scores for multiple gene sets and visualize across groups.
}
\examples{
obj <- SeuratVisProExample(
    n_cells = 300,
    n_genes = 1000,
    n_clusters = 10,
    seed = 123,
    genes_mt = "^MT-",
    neighbor_dims = 10,
    cluster_res = 0.5,
    umap_dims = 10,
    spatial = FALSE)

sets <- list(
  SetA = paste0('G', 1:10),
  SetB = paste0('G', 11:20))

res <- VisMetaFeature(
  obj,
  feature_sets = sets,
  group.by = "seurat_clusters",
  nbin = 24,
  min.size = 3,
  palette = "C",
  violin_width = 0.8,
  violin_alpha = 0.3,
  box_width = 0.3,
  box_alpha = 0.5)

res$plot

}
\author{
benben-miao
}
