% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisQCPanel.R
\name{VisQCPanel}
\alias{VisQCPanel}
\title{Interactive QC panel for Seurat v5 objects}
\usage{
VisQCPanel(
  object,
  genes_mt = "^MT-",
  genes_ribo = NULL,
  group.by = NULL,
  interactive = FALSE,
  assay = NULL,
  palette = "C"
)
}
\arguments{
\item{object}{A \code{Seurat} object.}

\item{genes_mt}{Regex for mitochondrial genes. Default \verb{^MT-} for human.}

\item{genes_ribo}{Regex for ribosomal genes (e.g., \verb{^RPL|^RPS}).}

\item{group.by}{Column in \code{object@meta.data} used for grouping.}

\item{interactive}{If \code{TRUE}, returns interactive plots via \code{plotly::ggplotly}.}

\item{assay}{Assay to use; defaults to \code{DefaultAssay(object)}.}

\item{palette}{Viridis palette option.}
}
\value{
A assembled ggplot or a list of interactive plots.
}
\description{
Visualize key QC metrics with customizable thresholds and optional interactivity.
}
\examples{
obj <- SeuratVisProExample()
p <- VisQCPanel(obj, genes_mt = "^MT-", genes_ribo = "^RPL|^RPS", group.by = "seurat_clusters", interactive = FALSE, palette = "C")
p

}
\author{
benben-miao
}
