% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisSpatialOverlay.R
\name{VisSpatialOverlay}
\alias{VisSpatialOverlay}
\title{Spatial feature overlay}
\usage{
VisSpatialOverlay(
  object,
  features,
  image = NULL,
  coords_cols = c("x", "y"),
  palette = "C",
  point_size = 2,
  alpha = 0.5
)
}
\arguments{
\item{object}{A \code{Seurat} object; if not spatial, falls back to embedding overlay.}

\item{features}{Character vector of gene features.}

\item{image}{Image key; defaults to active image.}

\item{coords_cols}{Column names for fallback coordinates.}

\item{palette}{Viridis palette option for color/fill.}

\item{point_size}{Point size for fallback scatter.}

\item{alpha}{Point alpha.}
}
\value{
A patchwork of spatial feature plots.
}
\description{
Overlay selected features onto spatial images with improved color scales.
}
\examples{
obj <- SeuratVisProExample(
    n_cells = 300,
    n_genes = 1000,
    n_clusters = 10,
    seed = 123,
    genes_mt = "^MT-",
    neighbor_dims = 10,
    cluster_res = 0.5,
    umap_dims = 10,
    spatial = TRUE)

p <- VisSpatialOverlay(
  obj,
  features = c("G1", "G2", "G3", "G4"),
  image = NULL,
  coords_cols = c("x", "y"),
  palette = "C",
  point_size = 2,
  alpha = 0.5)
p

}
\author{
benben-miao
}
